Updating to BigNumber for some things

+++ core/src/main/kotlin/com/dllewellyn/coinbaseapi/models/trade/CurrencyBuyAndSell.kt
@@ -1,3 +1,5 @@
 package com.dllewellyn.coinbaseapi.models.trade
 
-data class CurrencyBuyAndSell(val currencyFrom : String, val currencyTo : String, val buy : Double, val sell : Double)
+import java.math.BigInteger
+
+data class CurrencyBuyAndSell(val currencyFrom : String, val currencyTo : String, val buy : BigInteger, val sell : BigInteger)
